requirements.txt
streamlit~=1.33.0
python-dotenv~=1.0.1
openai~=1.16.2
langchain~=0.1.14
pandas~=2.2.1
requests~=2.31.0
docx2txt~=0.8
beautifulsoup4~=4.12.3
selenium~=4.19.0
anthropic~=0.23.1